:zap: perf: 권한 검사 추가, 소스코드 이동
