+++ src/main/java/com/exam/tomatoback/web/controller/chat/ChatApiController.java
@@ -1,6 +1,7 @@
 package com.exam.tomatoback.web.controller.chat;
 
 import com.exam.tomatoback.chat.service.ChatApiService;
+import com.exam.tomatoback.infrastructure.util.Constants;
 import com.exam.tomatoback.user.model.User;
 import com.exam.tomatoback.user.service.UserService;
 import com.exam.tomatoback.web.dto.chat.api.ChatListPageRequest;
@@ -10,13 +11,21 @@ import com.exam.tomatoback.web.dto.chat.api.ChatRoomResponse;
 import com.exam.tomatoback.web.dto.common.CommonResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.core.io.Resource;
+import org.springframework.core.io.UrlResource;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.net.MalformedURLException;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 
 @Controller
 @RestController
@@ -27,6 +36,7 @@ public class ChatApiController {
     private final UserService userService;
 
 
+    // login
     // 채팅 목록 불러오기
     @GetMapping
     public ResponseEntity<?> getChatList(@Valid ChatListPageRequest chatListPageRequest) {
@@ -34,12 +44,14 @@ public class ChatApiController {
     }
 
 
+    // login
     // 채팅룸에 있는 채팅 불러오기
     @GetMapping("/{roomId}")
     public ResponseEntity<?> getChat(@Valid ChatPageRequest chatPageRequest, @PathVariable long roomId) {
         return ResponseEntity.status(HttpStatus.OK).body(CommonResponse.success(chatApiService.getChats(chatPageRequest, roomId)));
     }
 
+    // login
     // 신규 채팅 룸 만들기
     @GetMapping("/room")
     public ResponseEntity<?> getChatRoom(ChatRoomRequest chatRoomRequest) {
@@ -50,6 +62,7 @@ public class ChatApiController {
         return ResponseEntity.status(HttpStatus.OK).body(CommonResponse.success(new ChatRoomResponse(chatRoomIdOrCreateRoomId, userByUserId.getId(), userByUserId.getNickname())));
     }
 
+    // login
     // 마지막 읽은 챗
     @GetMapping("/room/{roomId}/chat/{chatId}")
     public ResponseEntity<?> setLastRead(@PathVariable long roomId, @PathVariable long chatId) {
@@ -67,6 +80,36 @@ public class ChatApiController {
     @GetMapping("/user/{userId}")
     public ResponseEntity<?> getUserSellingList(@PathVariable long userId) {
         return ResponseEntity.status(HttpStatus.OK).body(CommonResponse.success(chatApiService.getUserSellingList(userId)));
+    }
+
+    @GetMapping("/images/{filename}")
+    public ResponseEntity<Resource> getImage(@PathVariable String filename) {
+        try {
+            Path filePath = Paths.get(Constants.CHAT_IMAGE_DIR).resolve(filename).normalize();
+            Resource resource = new UrlResource(filePath.toUri());
+
+            if (!resource.exists()) {
+                return ResponseEntity.notFound().build();
+            }
+
+            // 간단하게 이미지로 한정 (jpg/png/gif 등)
+            String contentType = "application/octet-stream";
+            if (filename.endsWith(".jpg") || filename.endsWith(".jpeg")) {
+                contentType = "image/jpeg";
+            } else if (filename.endsWith(".png")) {
+                contentType = "image/png";
+            } else if (filename.endsWith(".gif")) {
+                contentType = "image/gif";
+            }
+
+            return ResponseEntity.ok()
+                    .contentType(MediaType.parseMediaType(contentType))
+                    .header(HttpHeaders.CONTENT_DISPOSITION, "inline; filename=\"" + resource.getFilename() + "\"")
+                    .body(resource);
+
+        } catch (MalformedURLException e) {
+            return ResponseEntity.badRequest().build();
+        }
     }
